WifiDevice refactor
- Move svc impls to the WifiDevice itself
  - This way any stack can be uses once connection is handled
  - Added compat impl for Wifi which forwards to the WifiDevice impls
- Added internal wifi error enum and upgraded the basic wifi error enum
