+++ Cargo.toml
@@ -32,6 +32,8 @@ linked_list_allocator = { version = "0.10.3", default-features = false, features
 embedded-io = "0.3.1"
 fugit = "0.3.6"
 heapless = { version = "0.7.14", default-features = false }
+num-derive = { version = "0.3", features = ["full-syntax"] }
+num-traits = { version = "0.2", default-features = false }
 
 [build-dependencies]
 riscv-target = { version = "0.1.2", optional = true }
